* make config parser public
* do not create redundant tokio runtimes
